staging: bcm2835-camera: Match MMAL buffers to V4L2.
For historical reasons, the number of buffers passed to
the VPU over MMAL did not match that passed from V4L2.
That is a silly situation as the driver has to duplicate
serialisation and other functions that have already been
implemented in V4L2/videobuf2.
As we had more V4L2 buffers than MMAL ones, the MMAL
buffer headers were returned to the VPU immediately on
being filled, which is now invalid.

Match the number of buffers notified in queue_setup with
that used in MMAL.
Return buffers only when we get them from V4L2.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>