staging: bcm2835-camera: Allocate context once per buffer
The struct mmal_msg_context was being allocated for every
message being sent to the VPU, and freed when it came back.
Whilst that is required behaviour for some message (mainly
the synchronous ones), it is wasteful for the video buffers
that make up the majority of the traffic.

Add to the buffer_init/cleanup hooks that it allocates/frees
the msg_context required.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>