staging: vchiq-mmal: Support event messages
The VPU can generate events for things such as errors,
or resolution changes. These buffers do not come from the
normal buffer pools, and therefore a new message context
is added for each port.
There should never be a case of multiple events being raised
simultaneously, therefore dropping it should the buffer be in
use is a valid approach.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>