WIP: drm/vc4: Pass an update_id to the fw to return on completion
This should allow KMS to return the events correctly for
page flip completions.

ISSUE SEEN: On quiting X there appears to be a mode change
that is timing out because we aren't returning a completion.
I'm wondering if this is due to kicking off the mode change
when the update is still active, resulting in the fw never
giving us the appropriate callback.
DRM has a 10sec timeout which clears the issue, but it isn't
clean at present.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>