media: bcm2835-unicam: Remove any requirement to align the height
The requirement had been added to align the buffer size so that
the height was a multiple of 16 to match in with other hardware
acceleration APIs.
Most of the useful components that this is likely to be connected
to have been fixed to not require any specific alignment on the
height, therefore this restriction can be removed.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>