media: i2c: imx290: Fix the pixel rate at 148.5Mpix/s
Whilst the datasheet lists the link frequency changing between
1080p and 720p modes, reality is that with the default blanking
we have
(1920 + 280) * (1080 + 45) * 60fps = 148.5MPix/s
and
(1280 + 2020) * (720 + 30) * 60fps = 148.5MPix/s
and this reflects the readout rate whether in 10 or 12 bit
readout modes.

Measuring the link frequency, the CSI2 link frequencies
do follow the driver definitions, which implies there is a
FIFO between the pixel array and CSI2 transmitter.
The onlyregister differences between 1080p and 720p modes
are 0x305c and 0x305d so presumably these control the CSI2
block. Running with the 1080p settings works in 720p, but not
the other way around.

Report the pixel rate correctly regardless of link frequency
or bit depth.
Whilst we could switch to using the same link frequency for all
modes doing so would require complication to the driver as device
tree would still request both frequencies.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>