media: bcm2835-unicam: Add basic support for interlaced video
The ADV7282-M can produce interlaced video and signals the field
via the line number in the frame start and frame end short packets
being either 1 or 2.
Unicam can read this out via the packet capture handler. This is
already in use for frame ends due to the odd glitch in the
genuine frame end handling, therefore we already have the values
to use.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>