drm/bridge: ti-sn65dsi83: Move init code to pre_enable
The datasheet says that the initialisation should take
place with the DSI peripheral initialised, which means
the clock lane in HS, and the data lanes in LP-11.
Currently the initialisation happens in _enable, which
means the data lanes will be in HS mode too.

Now we have the option to pre_enable the upstream node
first, use that flag and move the initialisation to
_pre_enable.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>